+++ LMO_G9/model/Composer.cs
@@ -7,7 +7,7 @@ namespace LMO_G9.model
 {
     public class Composer : BaseModel
     {
-        private long composerId;
+        private int composerId;
         private string name;
         private string imagePath;
 
@@ -15,14 +15,14 @@ namespace LMO_G9.model
         {
         }
 
-        public Composer(long composerId, string name, string imagePath)
+        public Composer(int composerId, string name, string imagePath)
         {
             this.ComposerId = composerId;
             this.Name = name;
             this.ImagePath = imagePath;
         }
 
-        public long ComposerId { get => composerId; set => composerId = value; }
+        public int ComposerId { get => composerId; set => composerId = value; }
         public string Name { get => name; set => name = value; }
         public string ImagePath { get => imagePath; set => imagePath = value; }
     }
