+++ LMO_G9/model/MusicFactory.cs
@@ -7,23 +7,23 @@ namespace LMO_G9.model
 {
     public class MusicFactory : BaseModel
     {
-        private long musicFactoryId;
-        private long musicId;
-        private long accountId;
+        private int musicFactoryId;
+        private int musicId;
+        private int accountId;
 
         public MusicFactory()
         {
         }
 
-        public MusicFactory(long musicFactoryId, long musicId, long accountId)
+        public MusicFactory(int musicFactoryId, int musicId, int accountId)
         {
             this.MusicFactoryId = musicFactoryId;
             this.MusicId = musicId;
             this.AccountId = accountId;
         }
 
-        public long MusicFactoryId { get => musicFactoryId; set => musicFactoryId = value; }
-        public long MusicId { get => musicId; set => musicId = value; }
-        public long AccountId { get => accountId; set => accountId = value; }
+        public int MusicFactoryId { get => musicFactoryId; set => musicFactoryId = value; }
+        public int MusicId { get => musicId; set => musicId = value; }
+        public int AccountId { get => accountId; set => accountId = value; }
     }
 }
