+++ LMO_G9/model/MusicFavorite.cs
@@ -7,23 +7,23 @@ namespace LMO_G9.model
 {
     public class MusicFavorite : BaseModel
     {
-        private long musicFavoriteId;
-        private long musicId;
-        private long accountId;
+        private int musicFavoriteId;
+        private int musicId;
+        private int accountId;
 
         public MusicFavorite()
         {
         }
 
-        public MusicFavorite(long musicFavoriteId, long musicId, long accountId)
+        public MusicFavorite(int musicFavoriteId, int musicId, int accountId)
         {
             this.MusicFavoriteId = musicFavoriteId;
             this.MusicId = musicId;
             this.AccountId = accountId;
         }
 
-        public long MusicFavoriteId { get => musicFavoriteId; set => musicFavoriteId = value; }
-        public long MusicId { get => musicId; set => musicId = value; }
-        public long AccountId { get => accountId; set => accountId = value; }
+        public int MusicFavoriteId { get => musicFavoriteId; set => musicFavoriteId = value; }
+        public int MusicId { get => musicId; set => musicId = value; }
+        public int AccountId { get => accountId; set => accountId = value; }
     }
 }
