+++ LMO_G9/model/Singer.cs
@@ -7,7 +7,7 @@ namespace LMO_G9.model
 {
     public class Singer : BaseModel
     {
-        private long singerId;
+        private int singerId;
         private string name;
         private string imagePath;
 
@@ -15,14 +15,14 @@ namespace LMO_G9.model
         {
         }
 
-        public Singer(long singerId, string name, string imagePath)
+        public Singer(int singerId, string name, string imagePath)
         {
             this.SingerId = singerId;
             this.Name = name;
             this.ImagePath = imagePath;
         }
 
-        public long SingerId { get => singerId; set => singerId = value; }
+        public int SingerId { get => singerId; set => singerId = value; }
         public string Name { get => name; set => name = value; }
         public string ImagePath { get => imagePath; set => imagePath = value; }
     }
